requirements-dev.txt
# Development requirements
pyproj
matplotlib
cartopy>=0.18
pytest
pytest-cov
pytest-mpl
coverage
sphinx=2.2.1
sphinx_rtd_theme=0.4.3
sphinx-gallery
numpydoc
